autofocus and comments
username is now autofocused and debug comments are deleted. Caleb and I found that when "&& $hasErrors" is commented out in the if at the end of the form, the database recieved the info even when errors are thrown from processSignupForm.php. The problem lies within the validate function of the process form file.