logout works!
added if statements surrounding entire <html> tags so that if $_SESSION["valid"] == true, the page is displayed, otherwise there is an else statement that redirects to login. session_start() function is also called at the beginnign of every page (that requires a logged in user) so that each page has access to all session variables.